LPD-37932 rename util package to helper
