+++ modules/apps/commerce/commerce-frontend-api/bnd.bnd
@@ -2,5 +2,5 @@ Bundle-Name: Liferay Commerce Frontend API
 Bundle-SymbolicName: com.liferay.commerce.frontend.api
 Bundle-Version: 18.1.1
 Export-Package:\
-	com.liferay.commerce.frontend.model,\
-	com.liferay.commerce.frontend.util
+	com.liferay.commerce.frontend.helper,\
+	com.liferay.commerce.frontend.model
