+++ modules/apps/commerce/commerce-frontend-api/src/main/java/com/liferay/commerce/frontend/helper/ProductHelper.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.commerce.frontend.util;
+package com.liferay.commerce.frontend.helper;
 
 import aQute.bnd.annotation.ProviderType;
