+++ modules/apps/commerce/commerce-frontend-api/src/main/resources/com/liferay/commerce/frontend/helper/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
