--- modules/apps/commerce/commerce-frontend-api/src/main/resources/com/liferay/commerce/frontend/util/packageinfo
@@ -1,1 +1,0 @@
-version 3.1.0
