+++ modules/apps/commerce/commerce-frontend-taglib/src/main/resources/META-INF/resources/header/init.jsp
@@ -18,8 +18,8 @@ taglib uri="http://liferay.com/tld/theme" prefix="liferay-theme" %><%@
 taglib uri="http://liferay.com/tld/ui" prefix="liferay-ui" %>
 
 <%@ page import="com.liferay.commerce.constants.CommercePortletKeys" %><%@
+page import="com.liferay.commerce.frontend.helper.HeaderHelperUtil" %><%@
 page import="com.liferay.commerce.frontend.model.HeaderActionModel" %><%@
-page import="com.liferay.commerce.frontend.util.HeaderHelperUtil" %><%@
 page import="com.liferay.frontend.taglib.clay.servlet.taglib.util.DropdownItem" %><%@
 page import="com.liferay.petra.string.StringPool" %><%@
 page import="com.liferay.portal.kernel.language.LanguageUtil" %><%@
