+++ modules/apps/commerce/commerce-order-web/src/main/java/com/liferay/commerce/order/web/internal/portlet/action/EditCommerceOrderSummaryMVCRenderCommand.java
@@ -9,7 +9,7 @@ import com.liferay.commerce.configuration.CommerceOrderItemDecimalQuantityConfig
 import com.liferay.commerce.constants.CommercePortletKeys;
 import com.liferay.commerce.currency.util.CommercePriceFormatter;
 import com.liferay.commerce.exception.NoSuchOrderException;
-import com.liferay.commerce.frontend.util.CommerceOrderStepTrackerHelper;
+import com.liferay.commerce.frontend.helper.CommerceOrderStepTrackerHelper;
 import com.liferay.commerce.model.CommerceOrder;
 import com.liferay.commerce.notification.service.CommerceNotificationQueueEntryLocalService;
 import com.liferay.commerce.order.engine.CommerceOrderEngine;
